bug fixes lib/api
